Changed the suggrested API
So, it turns out that fixer.io now requires an API key, which in turn requires signup etc.
Since the point of the test is not to find out if candidates are willing to signup to
an API, but rather to see if they can read the docs, find a client library or something
similar, it seems prudent to change our suggestion of API so that it's simpler for candidates
to get the real purpose of the challenge.